+++ examples/Example102_StationaryConvectionDiffusion1D.jl
@@ -102,10 +102,8 @@ function calculate(grid,data,flux,verbose)
     enable_species!(sys,1,[1])
     
     ## Set boundary conditions
-    sys.boundary_values[1,1]=0.0
-    sys.boundary_values[1,2]=1.0
-    sys.boundary_factors[1,1]=VoronoiFVM.Dirichlet
-    sys.boundary_factors[1,2]=VoronoiFVM.Dirichlet
+    boundary_dirichlet!(sys,1,1,0.0)
+    boundary_dirichlet!(sys,1,2,1.0)
     
     ## Create a solution array
     inival=unknowns(sys)